refactor: timestamp module reorganization
