+++ crates/proof-of-sql-parser/src/intermediate_ast.rs
@@ -4,7 +4,9 @@
 * https://docs.rs/vervolg/latest/vervolg/ast/enum.Statement.html
 ***/
 
-use crate::{intermediate_decimal::IntermediateDecimal, parser_time::PoSQLTimestamp, Identifier};
+use crate::{
+    intermediate_decimal::IntermediateDecimal, posql_time::timestamp::PoSQLTimestamp, Identifier,
+};
 use serde::{Deserialize, Serialize};
 
 /// Representation of a SetExpression, a collection of rows, each having one or more columns.
