+++ crates/proof-of-sql-parser/src/lib.rs
@@ -3,7 +3,7 @@
 /// Module for handling an intermediate decimal type received from the lexer.
 pub mod intermediate_decimal;
 /// Module for handling an intermediate timestamp type received from the lexer.
-pub mod parser_time;
+pub mod posql_time;
 #[macro_use]
 extern crate lalrpop_util;
 
@@ -18,7 +18,8 @@ pub(crate) mod test_utility;
 pub(crate) mod select_statement;
 pub use select_statement::SelectStatement;
 
-pub(crate) mod error;
+/// Error definitions for proof-of-sql-parser
+pub mod error;
 pub use error::ParseError;
 pub(crate) use error::ParseResult;
