+++ crates/proof-of-sql-parser/src/posql_time/timezone.rs
@@ -1,0 +1,73 @@
+use crate::error::PoSQLTimestampError;
+use core::fmt;
+use serde::{Deserialize, Serialize};
+use std::sync::Arc;
+
+/// Captures a timezone from a timestamp query
+#[derive(Debug, Clone, Copy, Hash, Serialize, Deserialize, PartialEq, Eq)]
+pub enum PoSQLTimeZone {
+    /// Default variant for UTC timezone
+    Utc,
+    /// TImezone offset in seconds
+    FixedOffset(i32),
+}
+
+impl PoSQLTimeZone {
+    /// Parse a timezone from a count of seconds
+    pub fn from_offset(offset: i32) -> Self {
+        if offset == 0 {
+            PoSQLTimeZone::Utc
+        } else {
+            PoSQLTimeZone::FixedOffset(offset)
+        }
+    }
+}
+
+impl TryFrom<&Option<Arc<str>>> for PoSQLTimeZone {
+    type Error = PoSQLTimestampError;
+
+    fn try_from(value: &Option<Arc<str>>) -> Result<Self, Self::Error> {
+        match value {
+            Some(tz_str) => {
+                let tz = Arc::as_ref(tz_str).to_uppercase();
+                match tz.as_str() {
+                    "Z" | "UTC" | "00:00" | "+00:00" | "0:00" | "+0:00" => Ok(PoSQLTimeZone::Utc),
+                    tz if tz.chars().count() == 6
+                        && (tz.starts_with('+') || tz.starts_with('-')) =>
+                    {
+                        let sign = if tz.starts_with('-') { -1 } else { 1 };
+                        let hours = tz[1..3]
+                            .parse::<i32>()
+                            .map_err(|_| PoSQLTimestampError::InvalidTimezoneOffset)?;
+                        let minutes = tz[4..6]
+                            .parse::<i32>()
+                            .map_err(|_| PoSQLTimestampError::InvalidTimezoneOffset)?;
+                        let total_seconds = sign * ((hours * 3600) + (minutes * 60));
+                        Ok(PoSQLTimeZone::FixedOffset(total_seconds))
+                    }
+                    _ => Err(PoSQLTimestampError::InvalidTimezone(tz.to_string())),
+                }
+            }
+            None => Ok(PoSQLTimeZone::Utc),
+        }
+    }
+}
+
+impl fmt::Display for PoSQLTimeZone {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match *self {
+            PoSQLTimeZone::Utc => {
+                write!(f, "00:00")
+            }
+            PoSQLTimeZone::FixedOffset(seconds) => {
+                let hours = seconds / 3600;
+                let minutes = (seconds.abs() % 3600) / 60;
+                if seconds < 0 {
+                    write!(f, "-{:02}:{:02}", hours.abs(), minutes)
+                } else {
+                    write!(f, "+{:02}:{:02}", hours, minutes)
+                }
+            }
+        }
+    }
+}
