+++ crates/proof-of-sql-parser/src/posql_time/unit.rs
@@ -1,0 +1,49 @@
+use arrow::datatypes::TimeUnit as ArrowTimeUnit;
+use core::fmt;
+use serde::{Deserialize, Serialize};
+
+/// An intermediate type representing the time units from a parsed query
+#[derive(Debug, Clone, Copy, Hash, Serialize, Deserialize, PartialEq, Eq)]
+pub enum PoSQLTimeUnit {
+    /// Represents seconds with precision 0: ex "2024-06-20 12:34:56"
+    Second,
+    /// Represents milliseconds with precision 3: ex "2024-06-20 12:34:56.123"
+    Millisecond,
+    /// Represents microseconds with precision 6: ex "2024-06-20 12:34:56.123456"
+    Microsecond,
+    /// Represents nanoseconds with precision 9: ex "2024-06-20 12:34:56.123456789"
+    Nanosecond,
+}
+
+impl From<PoSQLTimeUnit> for ArrowTimeUnit {
+    fn from(unit: PoSQLTimeUnit) -> Self {
+        match unit {
+            PoSQLTimeUnit::Second => ArrowTimeUnit::Second,
+            PoSQLTimeUnit::Millisecond => ArrowTimeUnit::Millisecond,
+            PoSQLTimeUnit::Microsecond => ArrowTimeUnit::Microsecond,
+            PoSQLTimeUnit::Nanosecond => ArrowTimeUnit::Nanosecond,
+        }
+    }
+}
+
+impl From<ArrowTimeUnit> for PoSQLTimeUnit {
+    fn from(unit: ArrowTimeUnit) -> Self {
+        match unit {
+            ArrowTimeUnit::Second => PoSQLTimeUnit::Second,
+            ArrowTimeUnit::Millisecond => PoSQLTimeUnit::Millisecond,
+            ArrowTimeUnit::Microsecond => PoSQLTimeUnit::Microsecond,
+            ArrowTimeUnit::Nanosecond => PoSQLTimeUnit::Nanosecond,
+        }
+    }
+}
+
+impl fmt::Display for PoSQLTimeUnit {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            PoSQLTimeUnit::Second => write!(f, "seconds (precision: 0)"),
+            PoSQLTimeUnit::Millisecond => write!(f, "milliseconds (precision: 3)"),
+            PoSQLTimeUnit::Microsecond => write!(f, "microseconds (precision: 6)"),
+            PoSQLTimeUnit::Nanosecond => write!(f, "nanoseconds (precision: 9)"),
+        }
+    }
+}
