+++ crates/proof-of-sql/src/base/commitment/column_bounds.rs
@@ -290,7 +290,7 @@ mod tests {
     use super::*;
     use crate::base::{database::OwnedColumn, math::decimal::Precision, scalar::Curve25519Scalar};
     use itertools::Itertools;
-    use proof_of_sql_parser::parser_time::{PoSQLTimeUnit, PoSQLTimeZone};
+    use proof_of_sql_parser::posql_time::{timezone, unit::PoSQLTimeUnit};
 
     #[test]
     fn we_can_construct_bounds_by_method() {
@@ -533,7 +533,7 @@ mod tests {
 
         let timestamp_column = OwnedColumn::<Curve25519Scalar>::TimestampTZ(
             PoSQLTimeUnit::Second,
-            PoSQLTimeZone::Utc,
+            timezone::PoSQLTimeZone::Utc,
             vec![1_i64, 2, 3, 4],
         );
         let committable_timestamp_column = CommittableColumn::from(&timestamp_column);
