+++ crates/proof-of-sql/src/base/database/column.rs
@@ -6,7 +6,7 @@ use crate::base::{
 use arrow::datatypes::{DataType, Field, TimeUnit as ArrowTimeUnit};
 use bumpalo::Bump;
 use proof_of_sql_parser::{
-    parser_time::{PoSQLTimeUnit, PoSQLTimeZone},
+    posql_time::{timezone::PoSQLTimeZone, unit::PoSQLTimeUnit},
     Identifier,
 };
 use rayon::iter::{IntoParallelRefIterator, ParallelIterator};
