+++ crates/proof-of-sql/src/base/database/literal_value.rs
@@ -1,5 +1,5 @@
 use crate::base::{database::ColumnType, math::decimal::Precision, scalar::Scalar};
-use proof_of_sql_parser::parser_time::{PoSQLTimeUnit, PoSQLTimeZone};
+use proof_of_sql_parser::posql_time::{timezone::PoSQLTimeZone, unit::PoSQLTimeUnit};
 use serde::{Deserialize, Serialize};
 
 /// Represents a literal value.
